# api/schemas/solver.py
"""Solver schemas."""

from datetime import date, datetime
from typing import Any

from pydantic import BaseModel, ConfigDict, Field


class SolveRequest(BaseModel):
    """Schema for solve request."""

    org_id: str = Field(..., description="Organization ID")
    from_date: date = Field(..., description="Start date for schedule")
    to_date: date = Field(..., description="End date for schedule")
    mode: str = Field("strict", description="Solve mode: strict or relaxed")
    change_min: bool = Field(False, description="Enable change minimization")


class ViolationInfo(BaseModel):
    """Schema for constraint violation."""

    constraint_key: str
    message: str
    entities: list[str]
    severity: str


class FairnessMetrics(BaseModel):
    """Schema for fairness metrics."""

    stdev: float
    per_person_counts: dict[str, int]


class SolutionMetrics(BaseModel):
    """Schema for solution metrics."""

    hard_violations: int
    soft_score: float
    health_score: float
    solve_ms: float
    fairness: FairnessMetrics


class AssignmentInfo(BaseModel):
    """Schema for assignment information."""

    event_id: str
    person_id: str
    event_start: datetime
    event_type: str


class SolveResponse(BaseModel):
    """Schema for solve response."""

    solution_id: int = Field(..., description="Database ID of saved solution")
    metrics: SolutionMetrics
    assignment_count: int
    violations: list[ViolationInfo]
    message: str


class SolutionResponse(BaseModel):
    """Schema for solution response."""

    id: int
    org_id: str
    solve_ms: float
    hard_violations: int
    model_config = ConfigDict(from_attributes=True)

    soft_score: float
    health_score: float
    metrics: dict[str, Any] | None
    created_at: datetime
    assignment_count: int = 0


from api.schemas.common import ListResponse

SolutionList = ListResponse[SolutionResponse]


class ExportFormat(BaseModel):
    """Schema for export format request."""

    format: str = Field(..., description="Export format: json, csv, pdf, or ics")
    scope: str = Field("org", description="Export scope: org, person:{id}, or team:{id}")