# tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ExportFormat": {
        "description": "Schema for export format request.",
        "properties": {
          "format": {
            "description": "Export format: json, csv, pdf, or ics",
            "title": "Format",
            "type": "string"
          },
          "scope": {
            "default": "org",
            "description": "Export scope: org, person:{id}, or team:{id}",
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "format"
        ],
        "title": "ExportFormat",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListResponse_SolutionResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SolutionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ListResponse[SolutionResponse]",
        "type": "object"
      },
...
      "SolutionResponse": {
        "description": "Schema for solution response.",
        "properties": {
          "assignment_count": {
            "default": 0,
            "title": "Assignment Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "hard_violations": {
            "title": "Hard Violations",
            "type": "integer"
          },
          "health_score": {
            "title": "Health Score",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "soft_score": {
            "title": "Soft Score",
            "type": "number"
          },
          "solve_ms": {
            "title": "Solve Ms",
            "type": "number"
          }
        },
        "required": [
          "id",
          "org_id",
          "solve_ms",
          "hard_violations",
          "soft_score",
          "health_score",
          "metrics",
          "created_at"
        ],
        "title": "SolutionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered volunteer scheduling and sign-up management",
    "title": "SignUpFlow API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/solutions/": {
      "get": {
        "description": "List solutions with optional filters.",
        "operationId": "listSolutions",
        "parameters": [
          {
            "description": "Filter by organization ID",
            "in": "query",
            "name": "org_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by organization ID",
              "title": "Org Id"
            }
          },
          {
            "description": "Page size, max 200",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size, max 200",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Number of rows to skip",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of rows to skip",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_SolutionResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Solutions",
        "tags": [
          "solutions"
        ]
      },
      "post": {
        "description": "Create a manual solution record (for testing or external import).\nNote: This does not create assignments, just the solution metadata.",
        "operationId": "createManualSolution",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Solution Data",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Manual Solution",
        "tags": [
          "solutions"
        ]
      }
    },
    "/api/v1/solutions/{solution_id}": {
      "delete": {
        "description": "Delete solution and all assignments.",
        "operationId": "deleteSolution",
        "parameters": [
          {
            "in": "path",
            "name": "solution_id",
            "required": true,
            "schema": {
              "title": "Solution Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Solution",
        "tags": [
          "solutions"
        ]
      },
      "get": {
        "description": "Get solution by ID.",
        "operationId": "getSolution",
        "parameters": [
          {
            "in": "path",
            "name": "solution_id",
            "required": true,
            "schema": {
              "title": "Solution Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Solution",
        "tags": [
          "solutions"
        ]
      }
    },
    "/api/v1/solutions/{solution_id}/assignments": {
      "get": {
        "description": "Get all assignments for a solution.",
        "operationId": "getSolutionAssignments",
        "parameters": [
          {
            "in": "path",
            "name": "solution_id",
            "required": true,
            "schema": {
              "title": "Solution Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Solution Assignments",
        "tags": [
          "solutions"
        ]
      }
    },
    "/api/v1/solutions/{solution_id}/export": {
      "post": {
        "description": "Export solution in various formats (CSV, ICS, JSON).",
        "operationId": "exportSolution",
        "parameters": [
          {
            "in": "path",
            "name": "solution_id",
            "required": true,
            "schema": {
              "title": "Solution Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportFormat"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Solution",
        "tags": [
          "solutions"
        ]
      }
    },
...
  }
}